BGMList取得機能を外部に出す
